`connector` no longer depends on `sql`.
Other things still to be changed in this vein:
- `core` shouldn’t depend on `sql` (involves extricating SQL from views)
- `common` shouldn’t depend on `js` (move `Data => JS` to mongo
  connector)
- `connector` shouldn’t depend on `frontend` (need to get rid of old
  mongo so fs can expect QScript rather than LP)
