+++ CHANGELOG/revision.md
@@ -1,0 +1,1 @@
+- eliminate `connector` dependency on `sql`.
