+++ common/src/main/scala/quasar/fs/PathError.scala
