+++ connector/src/main/scala/quasar/fs/QueryFile.scala
@@ -18,12 +18,15 @@ package quasar.fs
 
 import quasar.Predef._
 import quasar._, Planner._, RenderTree.ops._
+import quasar.common.{PhaseResult, PhaseResults, PhaseResultT, PhaseResultW}
+import quasar.connector.CompileM
 import quasar.contrib.matryoshka._
 import quasar.contrib.pathy._
 import quasar.effect.LiftedOps
 import quasar.fp._
 import quasar.fp.ski._
 import quasar.fp.eitherT._
+import quasar.frontend.SemanticErrsT
 import quasar.qscript._
 
 import matryoshka._, Recursive.ops._, TraverseT.ops._
