+++ connector/src/test/scala/quasar/fs/FileSystemDefSpec.scala
@@ -17,8 +17,9 @@
 package quasar.fs.mount
 
 import quasar.Predef._
+import quasar.QuasarSpecification
+import quasar.connector.EnvironmentError
 import quasar.fp.ski.κ
-import quasar.{EnvironmentError, QuasarSpecification}
 import quasar.fs._
 
 import monocle.std.{disjunction => D}
