+++ core/src/main/scala/quasar/fs/mount/hierarchical.scala
@@ -17,7 +17,8 @@
 package quasar.fs.mount
 
 import quasar.Predef._
-import quasar.{LogicalPlan, PhaseResults}
+import quasar.LogicalPlan
+import quasar.common.PhaseResults
 import quasar.contrib.pathy._
 import quasar.effect._
 import quasar.fp._, free._, eitherT._
