+++ core/src/test/scala/quasar/fs/mount/ViewFileSystemSpec.scala
@@ -22,6 +22,7 @@ import quasar._, LogicalPlan._
 import quasar.contrib.pathy._
 import quasar.effect.{Failure, KeyValueStore, MonotonicSeq}
 import quasar.fp._, eitherT._
+import quasar.frontend.SemanticErrors
 import quasar.fs._, InMemory.InMemState
 import quasar.sql.{InnerJoin => _, _}, ExprArbitrary._
 import quasar.std._, IdentityLib.Squash, StdLib._, set._
