+++ it/src/test/scala/quasar/physical/filesystems.scala
@@ -17,7 +17,7 @@
 package quasar.physical
 
 import quasar.Predef._
-import quasar.{EnvironmentError, EnvErr}
+import quasar.connector.{EnvironmentError, EnvErr}
 import quasar.config.{CfgErr, ConfigError}
 import quasar.contrib.pathy._
 import quasar.effect.Failure
