+++ marklogic/src/main/scala/quasar/physical/marklogic/fs/queryfile.scala
@@ -19,7 +19,7 @@ package quasar.physical.marklogic.fs
 import quasar.Predef._
 import quasar.NameGenerator
 import quasar.{Data, LogicalPlan, Planner => QPlanner}
-import quasar.{PhaseResult, PhaseResults, PhaseResultT}
+import quasar.common.{PhaseResult, PhaseResults, PhaseResultT}
 import quasar.contrib.matryoshka._
 import quasar.contrib.pathy._
 import quasar.effect.MonotonicSeq
