+++ mongodb/src/main/scala/quasar/physical/mongodb/WorkflowExecutor.scala
@@ -17,10 +17,10 @@
 package quasar.physical.mongodb
 
 import quasar.Predef._
-import quasar.{EnvironmentError, EnvErrT}
-import quasar.{NameGenerator => QNameGenerator, _}
-import quasar.fp.kleisli._
+import quasar.{NameGenerator => QNameGenerator}
+import quasar.connector.{EnvironmentError, EnvErrT}
 import quasar.fp.eitherT._
+import quasar.fp.kleisli._
 import quasar.fs._
 import quasar.javascript._
 import quasar.physical.mongodb.execution._
