+++ mongodb/src/main/scala/quasar/physical/mongodb/plannerQScript.scala
@@ -19,6 +19,7 @@ package quasar.physical.mongodb
 import scala.Predef.$conforms
 import quasar.Predef._
 import quasar._, Planner._, Type.{Const => _, Coproduct => _, _}
+import quasar.common.{PhaseResult, PhaseResults, PhaseResultT}
 import quasar.contrib.matryoshka._
 import quasar.fp._
 import quasar.fp.ski._
