+++ sparkcore/src/main/scala/quasar/physical/sparkcore/fs/hdfs/package.scala
@@ -17,28 +17,23 @@
 package quasar.physical.sparkcore.fs
 
 import quasar.Predef._
-import quasar.fs._
-import quasar.fs.QueryFile.ResultHandle
-import quasar.fs.mount._, FileSystemDef._
-import quasar.physical.sparkcore.fs.{readfile => corereadfile}
-import quasar.physical.sparkcore.fs.{queryfile => corequeryfile}
-import quasar.physical.sparkcore.fs.hdfs.writefile.HdfsWriteCursor
+import quasar.connector.EnvironmentError
+import quasar.contrib.pathy._
 import quasar.effect._
-import quasar.fs.ReadFile.ReadHandle
-import quasar.fs.WriteFile.WriteHandle
 import quasar.fp.TaskRef
 import quasar.fp.free._
-import quasar.EnvironmentError
-import quasar.fs.mount.ConnectionUri
-import quasar.contrib.pathy._
+import quasar.fs._, QueryFile.ResultHandle, ReadFile.ReadHandle, WriteFile.WriteHandle
+import quasar.fs.mount._, FileSystemDef._
+import quasar.physical.sparkcore.fs.{queryfile => corequeryfile, readfile => corereadfile}
+import quasar.physical.sparkcore.fs.hdfs.writefile.HdfsWriteCursor
 
 import java.net.URI
 import scala.sys
 
-import org.apache.hadoop.fs.{FileSystem => HdfsFileSystem}
 import org.apache.hadoop.conf.Configuration
-import pathy.Path._
+import org.apache.hadoop.fs.{FileSystem => HdfsFileSystem}
 import org.apache.spark._
+import pathy.Path._
 import scalaz._, Scalaz._
 import scalaz.concurrent.Task
